Fixed bug: https://github.com/cferdinandi/web-app-starter-kit/issues/4
If first page in All Pages list in the admin dashboard was set to
"logged out only" user access, Admin would be redirected to front-end.
WordPress was grabbing first page in loop and using that one's
`wpwebapp_user_access_setting` as the status for the whole page. Added
`!is_admin()` to function to fix.